requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.0.5
djangorestframework==3.12.4
djangotoolbox==1.8.0
djongo==1.3.4
pycodestyle==2.7.0
PyJWT==2.1.0
pymongo==3.11.4
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
Werkzeug==1.0.1
